requirements.txt
argh==0.26.2
baton==1.0.1
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-appconf==1.0.3
django-baton==1.5.4
django-imagekit==4.0.2
djangorestframework==3.9.4
docker-py==1.9.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
hgicommon==1.2.0
hgijson==1.3.1
httplib2==0.13.0
idna==2.8
oauth2client==1.5.2
pathtools==0.1.2
pilkit==2.0
Pillow==6.1.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.5.3
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
watchdog==0.8.3
websocket-client==0.56.0
